+++ __tests__/functional/callbacks.spec.js
@@ -6,8 +6,8 @@ describe('Callbacks', () => {
     await page.goto(`${baseUrl}/callbacks.html`)
 
     const mouse = page.mouse
-    // move to the middle of the page
-    await mouse.move(1, 1, { steps: 100 })
+    // move to the beginning of the page
+    await mouse.move(1, 1, { steps: 10 })
     await mouse.down()
     await wait(100)
     // move 200px down and to the end of the page
@@ -19,12 +19,6 @@ describe('Callbacks', () => {
     const executesFn = await page.evaluate(() => {
       ds.Area.scroll(['right'], 1)
       return {
-        onDragStartCalls: window.onDragStartCalls,
-        onDragMoveCalls: window.onDragMoveCalls,
-        onElementSelectCalls: window.onElementSelectCalls,
-        onElementUnselectCalls: window.onElementUnselectCalls,
-        callbackCalls: window.callbackCalls,
-
         onPreDragStartCallsPS: window.pubsub.onPreDragStartCalls,
         onDragStartCallsPS: window.pubsub.onDragStartCalls,
         onDragMoveCallsPS: window.pubsub.onDragMoveCalls,
@@ -34,13 +28,6 @@ describe('Callbacks', () => {
         autoScrollCallsPS: window.pubsub.autoScrollCalls,
       }
     })
-
-    expect(executesFn.onDragStartCalls.length).toBe(1)
-    expect(executesFn.onDragMoveCalls.length).toBe(100)
-    expect(executesFn.onElementSelectCalls.length).toBe(2)
-    expect(executesFn.onElementUnselectCalls.length).toBe(0)
-    expect(executesFn.callbackCalls.length).toBe(1)
-    expect(executesFn.callbackCalls[0].elements.length).toBe(2)
 
     expect(executesFn.onPreDragStartCallsPS.length).toBe(1)
     expect(executesFn.onPreDragStartCallsPS[0].isDragging).toBe(false)