+++ src/modules/SelectableSet.js
@@ -6,66 +6,36 @@ import { toArray, handleElementPositionAttribute } from '../methods'
 
 export default class SelectableSet extends Set {
   /**
-   * @type {DSElements}
-   * @private
-   * */
-  _initElements
-  /**
-   * @type {string}
-   * @private
-   * */
-  _className
-  /**
-   * @type {string}
-   * @private
-   * */
-  _hoverClassName
-  /**
-   * @type {boolean}
-   * @private
-   * */
-  _useTransform
-  /**
-   * @type {boolean}
-   * @private
-   * */
-  _draggability
-
-  /**
    * @constructor SelectableSet
-   * @param {Object} p
-   * @param {DSInputElements} p.elements
-   * @param {DragSelect} p.DS
-   * @param {string} p.className
-   * @param {string} p.hoverClassName
-   * @param {boolean} p.useTransform
-   * @param {boolean} p.draggability
+   * @param {{DS:DragSelect}} obj
    * @ignore
    */
-  constructor({
-    elements,
-    className,
-    hoverClassName,
-    draggability,
-    useTransform,
-    DS,
-  }) {
+  constructor({ DS }) {
     super()
     this.DS = DS
-    this._initElements = toArray(elements)
-    this._className = className
-    this._hoverClassName = hoverClassName
-    this._useTransform = useTransform
-    this._draggability = draggability
-
     this.DS.subscribe('Interaction:init', this.init)
+    // @ts-ignore: @todo: update to typescript
+    this.DS.PubSub.subscribe('Settings:updated:selectables', () => {
+      this.clear()
+      this.init()
+    })
+    // @ts-ignore: @todo: update to typescript
+    this.DS.subscribe('Settings:updated:selectableClass', ({ settings }) => {
+      this.forEach((el) => {
+        el.classList.remove(settings['selectableClass:pre'])
+        el.classList.add(settings['selectableClass'])
+      })
+    })
   }
 
-  init = () => this._initElements.forEach((el) => this.add(el))
+  init = () =>
+    toArray(this.DS.stores.SettingsStore.s.selectables).forEach((el) =>
+      this.add(el)
+    )
 
   /** @param {DSElement} element */
   add(element) {
-    element.classList.add(this._className)
+    element.classList.add(this.DS.stores.SettingsStore.s.selectableClass)
     element.addEventListener('click', this._onClick)
     element.addEventListener('mousedown', this._onPointer)
     element.addEventListener('touchstart', this._onPointer, {
@@ -73,7 +43,10 @@ export default class SelectableSet extends Set {
       passive: false,
     })
 
-    if (this._draggability && !this._useTransform)
+    if (
+      this.DS.stores.SettingsStore.s.draggability &&
+      !this.DS.stores.SettingsStore.s.useTransform
+    )
       handleElementPositionAttribute({
         computedStyle: window.getComputedStyle(element),
         node: element,
@@ -84,8 +57,8 @@ export default class SelectableSet extends Set {
 
   /** @param {DSElement} element */
   delete(element) {
-    element.classList.remove(this._className)
-    element.classList.remove(this._hoverClassName)
+    element.classList.remove(this.DS.stores.SettingsStore.s.selectableClass)
+    element.classList.remove(this.DS.stores.SettingsStore.s.hoverClass)
     element.removeEventListener('click', this._onClick)
     element.removeEventListener('mousedown', this._onPointer)
     element.removeEventListener('touchstart', this._onPointer, {
@@ -97,8 +70,10 @@ export default class SelectableSet extends Set {
 
   clear = () => this.forEach((el) => this.delete(el))
 
-  _onClick = (event) => this.DS.publish(['Selectable:click:pre', 'Selectable:click'], { event })
-  _onPointer = (event) => this.DS.publish(['Selectable:pointer:pre', 'Selectable:pointer'], { event })
+  _onClick = (event) =>
+    this.DS.publish(['Selectable:click:pre', 'Selectable:click'], { event })
+  _onPointer = (event) =>
+    this.DS.publish(['Selectable:pointer:pre', 'Selectable:pointer'], { event })
 
   /** @param {DSElements} elements */
   addAll = (elements) => elements.forEach((el) => this.add(el))