+++ src/stores/ScrollStore.js
@@ -13,10 +13,6 @@ export default class ScrollStore {
    * @private */
   _currentVal
   /**
-   * @type {DSArea}
-   * @private */
-  _areaElement
-  /**
    * @type {boolean}
    * @private */
   _canScroll
@@ -24,30 +20,34 @@ export default class ScrollStore {
   /**
    * @class ScrollStore
    * @constructor ScrollStore
-   * @param {{ DS:DragSelect, areaElement: DSArea, zoom:number }} p
+   * @param {{ DS:DragSelect }} p
    * @ignore
    */
-  constructor({ DS, areaElement, zoom }) {
-    this._areaElement = areaElement
+  constructor({ DS }) {
     this.DS = DS
-    this.zoom = zoom
-
     this.DS.subscribe('Interaction:init', this.init)
     this.DS.subscribe('Interaction:start', () => this.start())
     this.DS.subscribe('Interaction:end', () => this.reset())
   }
 
-  init = () => this._areaElement.addEventListener('scroll', this.update)
+  init = () =>
+    this.DS.stores.SettingsStore.s.area.addEventListener('scroll', this.update)
 
   start = () => {
-    this._currentVal = this._initialVal = getCurrentScroll(this._areaElement)
-    this._areaElement.addEventListener('scroll', this.update)
+    this._currentVal = this._initialVal = getCurrentScroll(
+      this.DS.stores.SettingsStore.s.area
+    )
+    this.DS.stores.SettingsStore.s.area.addEventListener('scroll', this.update)
   }
 
-  update = () => (this._currentVal = getCurrentScroll(this._areaElement))
+  update = () =>
+    (this._currentVal = getCurrentScroll(this.DS.stores.SettingsStore.s.area))
 
   stop = () => {
-    this._areaElement.removeEventListener('scroll', this.update)
+    this.DS.stores.SettingsStore.s.area.removeEventListener(
+      'scroll',
+      this.update
+    )
     this._initialVal = { x: 0, y: 0 }
     this._canScroll = null
   }
@@ -59,14 +59,14 @@ export default class ScrollStore {
 
   get canScroll() {
     if (typeof this._canScroll === 'boolean') return this._canScroll
-    return (this._canScroll = canScroll(this._areaElement))
+    return (this._canScroll = canScroll(this.DS.stores.SettingsStore.s.area))
   }
 
   get scrollAmount() {
     const scrollDiff = vect2.calc(this.currentVal, '-', this.initialVal)
 
     // if area is zoomed, the scroll values are skewed, we need to fix that manually :(
-    const zoom = vect2.num2vect(this.zoom)
+    const zoom = vect2.num2vect(this.DS.stores.SettingsStore.s.zoom)
     const zoomScroll = vect2.calc(
       vect2.calc(scrollDiff, '*', zoom),
       '-',
@@ -86,7 +86,7 @@ export default class ScrollStore {
 
   get currentVal() {
     if (!this._currentVal)
-      this._currentVal = getCurrentScroll(this._areaElement)
+      this._currentVal = getCurrentScroll(this.DS.stores.SettingsStore.s.area)
     return this._currentVal
   }
 }